+++ mlprogram/utils/data/__init__.py
@@ -1,3 +1,4 @@
 from mlprogram.utils.data.utils import ListDataset, to_map_style_dataset, transform  # noqa
 from mlprogram.utils.data.functions import get_words, get_characters, get_samples  # noqa
 from mlprogram.utils.data.functions import CollateOptions, Collate  # noqa
+from mlprogram.utils.data.random import random_split  # noqa
