+++ mlprogram/utils/random.py
@@ -1,6 +1,5 @@
-import torch
 import numpy as np
-from typing import List, Dict
+from typing import List
 from mlprogram import logging
 
 logger = logging.Logger(__name__)
@@ -19,19 +18,3 @@ def split(rng: np.random.RandomState, n: int, k: int, eps: float) -> List[int]:
         last = divisor
     ks.append(n - last)
     return ks
-
-
-def random_split(dataset: torch.utils.data.Dataset, ratio: Dict[str, float],
-                 seed: int) \
-        -> Dict[str, torch.utils.data.Dataset]:
-    logger.info(f"Split dataset with seed={seed}")
-    keys = list(ratio.keys())
-    total = len(dataset)
-    lengths = {key: int(total * ratio[key]) for key in keys}
-    remain = total - sum([value for value in lengths.values()])
-    lengths[keys[0]] += remain
-    datasets = torch.utils.data.random_split(
-        dataset,
-        [lengths[key] for key in keys],
-        generator=torch.Generator().manual_seed(seed))
-    return {key: datasets[i] for i, key in enumerate(keys)}
