+++ test/utils/test_random.py
@@ -1,22 +1,12 @@
 import unittest
-import torch
-from mlprogram.utils.random import random_split
+import numpy as np
+from mlprogram.utils.random import split
 
 
-class TestRandomSplit(unittest.TestCase):
+class TestSplit(unittest.TestCase):
     def test_happy_path(self):
-        dataset = torch.utils.data.TensorDataset(torch.rand(10,))
-        datasets = random_split(dataset, {"0": 0.5, "1": 0.5}, 0)
-        self.assertEqual(2, len(datasets))
-        self.assertEqual(5, len(datasets["0"]))
-        self.assertEqual(5, len(datasets["1"]))
-
-    def test_reamain(self):
-        dataset = torch.utils.data.TensorDataset(torch.rand(11,))
-        datasets = random_split(dataset, {"0": 0.5, "1": 0.5}, 0)
-        self.assertEqual(2, len(datasets))
-        self.assertEqual(6, len(datasets["0"]))
-        self.assertEqual(5, len(datasets["1"]))
+        ns = split(np.random.RandomState(0), 10, 3, 1e-5)
+        self.assertEqual(10, sum(ns))
 
 
 if __name__ == "__main__":
